legacy/src/Gateway/requirements.txt
cycler==0.12.1
fonttools==4.55.3
iniconfig==2.0.0
joblib==1.4.2
kiwisolver==1.4.7
numpy
nvidia-nccl-cu12==2.24.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==5.24.1
pluggy==1.5.0
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.6.0
scipy==1.7.3
six==1.17.0
tenacity==9.0.0
threadpoolctl==3.5.0
tzdata==2024.2
xgboost==2.1.3
flask==2.3.3
streamlit
requests
fastapi
uvicorn
flask-jwt-extended==4.4.0
python-jose

